# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ResPlan REST API",
    "description": "Календарно-сетевое и ресурсное планирование проектов: варианты использования UC-1 – UC-9",
    "version": "0.9.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/trainings": {
      "post": {
        "tags": [
          "UC-9 Резервировать время на обучение"
        ],
        "summary": "Добавить блок обучения в календарь сотрудника",
        "operationId": "reserve",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResultDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/trainings/{id}": {
      "delete": {
        "tags": [
          "UC-9 Резервировать время на обучение"
        ],
        "summary": "Отменить блок обучения",
        "operationId": "cancel_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "TrainingRequest": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "integer",
            "format": "int32"
          },
          "courseTitle": {
            "type": "string",
            "maxLength": 150,
            "minLength": 0
          },
          "provider": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "loadPercent": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 1
          },
          "protectedTime": {
            "type": "boolean"
          }
        },
        "required": [
          "courseTitle",
          "employeeId",
          "endDate",
          "loadPercent",
          "protectedTime",
          "startDate"
        ]
      },
      "BookingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "employeeId": {
            "type": "integer",
            "format": "int32"
          },
          "employee": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "projectId": {
            "type": "integer",
            "format": "int32"
          },
          "projectPriority": {
            "type": "integer",
            "format": "int32"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "loadPercent": {
            "type": "integer",
            "format": "int32"
          },
          "protectedTime": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "BookingResultDto": {
        "type": "object",
        "properties": {
          "booking": {
            "$ref": "#/components/schemas/BookingDto"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "conflicts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConflictDto"
            }
          }
        }
      },
      "ConflictDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "employee": {
            "type": "string"
          },
          "bookingA": {
            "type": "string"
          },
          "bookingB": {
            "type": "string"
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "totalLoad": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "first": {
            "$ref": "#/components/schemas/BookingDto"
          },
          "second": {
            "$ref": "#/components/schemas/BookingDto"
          },
          "detectedAt": {
            "type": "string",
            "format": "date-time"
          },
          "resolutionNote": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}